configs: add missing test config for aws-*-fips variants (#397)
The test harness for AWS now requires test configuration for all
variants to be present. The nightly publish run failed because the
aws-*-fips variants did not have any yet. Fix this by copying the test
config of the aws-base variant (the same config as any other AWS
variant).

See:
https://github.com/chainguard-dev/wolfi-vm/actions/runs/15455625873/job/43507267254

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>