auth/github: align github token with libraries
Rather than passing the github token in the package directory, follow
the same pattern as use for Chainguard libraries access and use
the melange cache directory to pass the token. This avoids accidentally
checking it into the git repo.

Passing of secrets to melange is currently in general discussion/design
so this is not the final solution, but improves the current situation.