enable hypervisor tests where platforms support it
AWS doesn't supported nested virt, they just offer bare metal VMs instead (we have no bare metal tests, so skipping that for now)

Azure supports on D series and E series VMs

GCP supports on Intel Haswell or later, with the right flag.

For qemu, it's totally dependent on what environment is running it so I opted not to add it.