amazon-eks-ami{,-fips}: unpin the dependency on cloud-provider-aws
The Kubernetes AWS Cloud Provider only targets compatibility with a matching
Kubernetes version. Therefore, unpin the dependency on a particular version of
the cloud-provider-aws package and leave it up to consumers of the package to
pin it to the desired version (or let it default to the latest available
version of Kubernetes). This avoids accidentally building Frankenstein's EKS
AMI mixing components targeting different versions of Kubernetes.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>