azure: Update Marketplace Technical Configuration (#390)
After some investigation I've decided to use this small set of changes
to the partnercenter-cli-extensions. partnercenter-cli-extensions are
still in preview and the changes I've made[1] haven't been reviewed.

They're pretty straightforward and we are currently only interested in
appending vm image versions for now. There are a few alternative
libraries and there is some dicsussion and examples in 10427 [2]

The main advantage of the partnercenter-cli-extension is that it uses
the auth infrastructure included in the broader az cli tooling. Most
libraries and scripts I've seen are using the same kind of API tokens
that we previously removed from our actions.

I included the built wheel here because It's easy for testing. We need a
wheel to install the extensionand I'm not aware of anywhere we can put
arbitrary assest. We could host the wheel in the release section of a
github repo as the az extension docs suggests or vendoring the code in
this repo and build the wheel as part of the hook.

The plan is to verify that the hooks work for pushing drafts my testing
plans and offers. Once we verify it works for those we can switch over
to the chainguard_eap offers and enable the preview submission.

A different PR that adds an action to make the offer live to the
marketplace will be drafted.

I'm using a different make target and script because there's already a
lot of options in the exiting one and marketplace and the azure
galleries are different infrastructure. Marketplace offers also need to
have consistent image states. If we have have our offer set to have
images for specific architecutres and generations we will be allowed to
upload them individually but we will fail the validation checks and be
unable to publish our offer. To prevent that these scripts assume that
both archtectures exist and if they don't the Marketplace API will
reject us for registering an image that doesn't providng a small check
to ensure our offer is always
publishable. In this way the Partnercenter portal is different than the
API. The portal does more comprehensive
checks and will not allow you to save a draft without sufficient image
coverage.

Example run:

```
~/chainguard-dev/wolfi-vm jvreeland-azure-marketplace-preview* ≡
❯ az partnercenter marketplace offer plan technical-configuration show --offer-id jvreeland_test_offer --plan-id jvreeland_test_plan | jq '.vmImageVersions[].versionNumber'
WARNING: Command group 'partnercenter marketplace offer plan technical-configuration' is in preview and under development. Reference and support levels: https://aka.ms/CLI_refstatus
"20250514.0018.0"

~/chainguard-dev/wolfi-vm jvreeland-azure-marketplace-preview* ≡
❯ make azure-marketplace-update-technical-plan-eap-dev PREFIX=chainguard AZVERSION=20250510.0018.0 AZGALLERY=chainguard_vms_eap AZMARKETPLACE_PLAN=jvreeland_test_plan AZMARKETPLACE_OFFER=jvreeland_test_offer &> /dev/null

~/chainguard-dev/wolfi-vm jvreeland-azure-marketplace-preview* ≡ 22s
❯ az partnercenter marketplace offer plan technical-configuration show --offer-id jvreeland_test_offer --plan-id jvreeland_test_plan | jq '.vmImageVersions[].versionNumber'
WARNING: Command group 'partnercenter marketplace offer plan technical-configuration' is in preview and under development. Reference and support levels: https://aka.ms/CLI_refstatus
"20250510.0018.0"
"20250514.0018.0"

~/chainguard-dev/wolfi-vm jvreeland-azure-marketplace-preview* ≡
❯ cat output/azure-marketplace-update-technical-plan-eap-devs.json
{
  "message": "completed",
  "result": "succeeded"
}
```

I'm not sure if everyone reviewing this has access to partner center if
you don't want want to test this out I can get you access or demo it on
a call.

1. https://github.com/Azure/partnercenter-cli-extension/pull/242/files
2. https://github.com/chainguard-dev/internal-dev/issues/10427

---------

Co-authored-by: Justin Vreeland <justin.vreeland@chainguard.com>