ct-manylinux-2.28: resolve conflicting stdatomic.h
Turns out there are two different atdatomic.h headers for C and C++
and each should be used for one or the other compiler. Meaning they
have to be in separate paths, and different paths should be used for C
and C++ compilers. Previously the assembled headers copied over each
other, and the C version of the header was not provided.

Also existing C++ headers cannot be used, as their contents is
different for old glibc.

Add test case of configuring ffmpeg which currently fails to find
atomics.