azure:  Fixup timestamp related failures. (#420)
Initially this was just a step in the previous job and would've gotten
the timestamp from the env of that job. When I split it out I didn't
preserve the timestamp env.

For the upload script I didn't keep the set -e I tested with and also
didn't add the helper from previous scripts to check failures. This adds
the helper function so we get a proper failure.

Co-authored-by: Justin Vreeland <justin.vreeland@chainguard.com>