Add variants for Amazon EKS 1.33 (#425)
Amazon EKS 1.33 was released on May 30. Add the new image variants
aws-eks-1.33-dev and aws-eks-1.33-fips-dev targeting the new release.

The new variants mirror aws-eks-1.32-dev and aws-eks-1.32-fips-dev
(formerly just aws-eks-dev and aws-eks-fips-dev) with exception for
dependencies tied to a version of Kubernetes. At the moment, these are
Kubelet and the Amazon ECR
credential helper.

At the moment, AMIs for the new variants will only be built but not yet
published to the AWS Marketplace since we do not have a product listing
for them yet. Creating one requires an AMI to be built first.

---------

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>