enable gcp tests where appropriate
Run the gcp routing tests in images with agents, and use a non-root user to test the guest agent's user creation and SSH flow.

startupscript tests are broken due to issues with the agent that can't be immediately resolved, I will follow up on that once I've fixed the agent.