google-guest-agent: change to manual updates, withdraw unsupported versions
Not all guest-agent tags are support releases. Not all guest-agent github releases are supported releases.

Drop the dependency on google-guest-agent-manager because that package a similar problem but we can't use any tags from there. This package is still fully functional without it.

We need to switch back to an older tag that's a supported release, now we need to withdraw all the unsupported ones we built and published, and bump epoch to rebuild with packaging changes.