Drop google-guest-agent-manager (#437)
This package has no tagged versions so we don't know what commits are OK
to build and release, and google-guest-agent is functional without it,
it just doesn't have some of the new features yet. We don't need this
package. See http://go/google-guest-agent-tag-problem for more info
about problem with google guest agent packages in general.