ct-manylinux-2.28: neutralise sys/single_threaded.h
glibc 2.32 added sys/single_threaded.h (and __libc_single_threaded
symbol) which some applications use for optimisations. The header is
discovered, and so is the symbol in the system libc, but it is not
available in the oldglibc.

Intentially replace the header with a c pre-processor error, such that
autoconf fails to detecting it as an available header and
automatically deconfiguring using it. This avoids manually overriding
each affected autoconf project with an argument
ac_cv_header_sys_single_threaded_h='no'.