add job to test vm in a aws eks cluster (#452)
This will add the job to run this manually, and allow input of
cluster name and region and ami.

The 'test-eks' helper script will not re-create a cluster if it
is present.  That is intentional, to re-use the cluster as it
takes 15 minutes to start a cluster, and it can fail with
unrelated options.

So while you _can_ change the cluster name and the region
generally speaking we don't want to.

In the future, we'll fix this to run more automatically and
do the right thing.  Walk first.

Co-authored-by: Scott Moser <smoser@brickies.net>