ct-manylinux-2.28: fix libm.so on aarch64
Turns out 2.28 libm.so on aarch64 is a regular library, and not a
linker script of libm + libvec like it is on x86_64.

Thus add special handling of libm linker scripts on x86_64, and
regular shared library landling on aarch64.

Including system overrides, which are always linker scripts to match
up to date system glibc-dev.

This fixes `cargo` usage on aarch64 on chainguard-2.28, and progresses
with bootstrapping rust-1.87.