tools: add aws-restrict-old-versions (#462)
Add a new aws-restrict-old-versions script to help restrict old product
versions on the AWS Marketplace. As there is a limit of 100 versions per
product after which no new versions can be uploaded, regularly
restricting versions becomes a necessity.

The script mostly automates this, but keeps a human in the loop for
review for now. Enhancements can be made to eventually have the script
operate without intervention and on a schedule. For now, we need the
human to take further steps anyway, though, and are figuring out those
details.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>