aws-eks-*-fips-dev: use FIPS build of aws-iam-authenticator (#468)
Use the new `aws-iam-authenticator-fips` package for the EKS FIPS
variants to get the FIPS build of AWS IAM Authenticator. A scan of
binaries on the image showed this to be the only package to be missing.
(I'm looking to integrate the scan into our automated tests.)

See: https://github.com/chainguard-dev/internal-dev/issues/11302

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>