nvidia-cuda-nvcc-12.9: Add package that automatically uses a supported gcc
Today every package that builds with CUDA needs to know which version(s)
of gcc are compatible with that version of nvcc. They needs to make sure
that gcc package is installed, and configure nvcc to use it (by installing
a gcc-X-default package, adding -ccbin to NVCC_APPEND_FLAGS, etc). This
also results in an extra diff between pkg-cuda-X.yaml and pkg-cuda-Y.yaml
files, which we'd like to minimize to help in future templating.

Here we addd a convenience package that depends on a known compatible
gcc version, and installs wrappers around it in $CUDA_HOME/bin.
nvcc looks for a compiler here before falling back to the $PATH.

This is an additional package because we may not want to force this
behavior for everyone. There's an off-chance that someone is counting
on nvcc to default to a different version of gcc that they can select
by installing gcc-X-default. It also let's us migrate packages one-by-one.

Note: We use wrappers instead of a symlinks because the existing
/usr/local/bin/gcc* wrappers installed by openssf-compiler-wrappers
choose which /usr/bin/gcc* to call based on "$0". So if
/usr/local/cuda-12.9/bin/gcc is a symlink to /usr/local/bin/gcc-14,
it will end up calling /usr/bin/gcc instead of /usr/bin/gcc-14.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>
[v2: install wrapper in contextdir, thx @javacruft]