Limit aws publication concurrency to 4 per make invocation. (#478)
There is an AWS limit on our publishing account that enforces 20
concurrent ImportSnapshot operations. The failure ends up with a message
like:

> botocore.exceptions.ClientError: An error occurred
> (ResourceCountLimitExceeded) when calling the ImportSnapshot
operation:
> Task failed to initialize - task limit (20) exceeded.

We have requested increase from 20 to 50 and when that goes through we
can increase this limit.

The number 4 was chosen because we currently have 2x2 matrix
(x86_64/aarch64 x eks/non-eks) giving us 4 concurrent actions. If we
keep each of those actions to 4 tasks at once we should not hit the
limit of 20.

The limit on the publishing (aws-publish) is probably not needed, but
certainly not harmful. The successful run takes 5 seconds or so.