ct-manylinux-2.28: statically link libgcc and libstdc++
pypa/manylinux upstream images use RHEL gcc-14-toolset that mixes
gcc-8 libstdc++ with backported symbols from gcc-14 to static
non-shared libstdc++. Such combo is not possible to trivially build
and relies on very invasive patches.

One option is to use a custom auditwheel policy to always vendor in
our copy of libstdc++.

Another option is to statically link libstdc++.

It is not obvious which path is best of the three. However statically
linking libstdc++ is likely the most portable and safe.

Note this was already applied in the wheel rebuilder project at:
- https://github.com/chainguard-dev/ecosystems-wheel-rebuilder/pull/398

But this is also needed during regular builds to ensure that vendored
shared libraries also do not have a high libstdc++ shared linking abi.