py3-tensorrt-cuda-12: Disable melange sca dependencies
These packages get automatic sca dependencies on so:libnvinfer=10 and
friends. Both tensorrt-10.7-cuda-12.6 and tensorrt-cuda-12 provides
packaages that supply them, so apk will happily use a version streamed
one instead of the mainline one. We want the mainline one, so we need
to turn of sca-generated dependencies and hardcode them instead.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>