Change from ignore-in-shared-libs to allow-shlib-undefined
Turns out unresolved-symbols=ignore-in-shared-libs is not the same as
allow-shlib-undefined.

The latter in the man-page says "It does not  affect  how  undefined  symbols  in  regular object files are handled." Which one would guess means report-all as that is stated as "This is the default" under `--unresolved-symbols` option.

Well that's not quite true. The default setting for object-files and shared-libs is actually unset, which is then converted to ignore-all in case `-shared` flag is passed, or set to "report-all" if it was not.

Coincedentally "ignore-in-object-files" changes values from unset of
both object-files and shared-libs which breaks the -shared flag doing
its magic, and there are projects that intentionally or accidentally
underlink things, and it happens to work as something else happens to
link things too. util-linux and libpam linking is prime example of
this, which is missing in autoconf build system, but is present in
meson build system.

Separately for ct-manylinux-2.28 we really want ignore-in-shared-libs
as during ABI downgrades it is ok to mix linking high and low abi
glibc transitive dependencies. But keeping the -shared flag
side-effects, thus we really must use `--allow-shlib-undefined` flag
instead.

This also explains why "ignore-all" setting didn't work correctly
either, because that was too much ignoring, including when -shared was
not passed and one was creating binaries that really should not
ignore-all.

References:
- https://github.com/bminor/binutils-gdb/blob/cf729dbef8b0a97aae040b9d9d4926d462001997/ld/lexsup.c#L1129
- https://github.com/bminor/binutils-gdb/blob/cf729dbef8b0a97aae040b9d9d4926d462001997/ld/lexsup.c#L1361
- https://github.com/chainguard-dev/enterprise-packages/pull/24587
- https://github.com/chainguard-dev/enterprise-packages/pull/24662

Attempt number 5 of getting this right.