Allow installing CUDA library stubs w/o adding to system path
I'd added /etc/ld.so.conf.d snippets to the packages that provide
the libcuda.so and libnvidia-ml.so stubs so that they'd be easier
to consume as build-time and test-time dependencies. But I failed
to consider that we ship these stubs to customers in non-devel images
as well (such as pytorch) and we don't want them to take precedence
over their GPU-host-provided libraries.

Let's revert that change from the packages we ship in images, and
create new metapackages for our internal use to serve the purpose.
That also allows us to use names that are easier to remember because
they include they contain the name of the needed stub:

 - libcuda-stub-cuda-XX-dev
 - libnvidia-ml-stub-cuda-XX-dev

Fixes: https://github.com/chainguard-dev/internal-dev/issues/14495

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>