claude - update to 1.0.43 and improve install script to ~/.claude (#2931)
Running 'claude' installed by this package would fail as it
wanted to update /usr/local and non-root user woudl not have
permissions to do so.

Running /migrate-installer
was an option to get that installer moved from the global install
dir to per-user.  And it would install to with
npm --prefix=$HOME/.claude/local/

So, that is what we do here.

First time running /usr/bin/claude, it will install the expected
version to ~/.claude/local/claude and then execute that.

Second time running, it will just exec ~/.claude/local/claude
straight away.