tritonserver: begin moving stack to CUDA 12.9 (#26307)
* tritonserver: Establish initial CUDA 12.9 builds

**This is not an attempt to maintain multiple CUDA variants of this stack.**

We currently can get into a situation where we have different parts of
the tritonserver package built for different versions of CUDA. When
this happens, images blow up in size because they now ship 2 different
CUDA runtimes. Worse, only one of these runtimes can be configured to
run on hardware at a time. Let's allow images to clearly pull in the
packages for exactly one stack. Once images have transitioned to the
new stack, the old stack can then be deleted.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* triton-inference-server: Package source repos for use as build-deps

tritonserver and its backends pull source from these repos at
build time without verifying the expected commit. When I noticed
this, I started a pattern of pre-fetching these in a `git-checkout`
pipeline (commit 32dcb3526ac070e3680fe81c67a3a50446947057).

The problem with that is we need to manually bump them forward. To avoid
that, here we make a package per repo so that our bots can keep them up to
date[*]. Of course, this means that a new repo tag could make an rdep ftbfs
that wasn't ready for it yet. But those rdeps should automatically be moving
forward as well. In the event that we're unable to move them forward
(unrelated build failure, for example), we can always pin the apk dependency
temporarily.

[*] Note that this will require adding support for tracking release branches
    to our infra. I've prototyped this support in the lifecycle tooling, but
    that will take some time to finish off and land.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* tritonserver-cuda-12.9: bump to latest upstream

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* py3-torch-tensorrt-cuda-12.9: bump to latest upstream

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* py3-torch-tensorrt-cuda-12.9: Use mainline pytorch

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* py3-torch-tensorrt-cuda-12.9: use system nvtx

Fixes a ftbfs w/ CUDA 12.9.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* py3-torch-tensorrt-cuda-12.9: Use TORCH_CUDA_ARCH_LIST from py3-torch-cuda-12.9

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* tritonserver-cuda-12.9: Bump to 2.59.0

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* tritonserver-backend-python-cuda-12.9: pin to cmake-3

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* tritonserver-cuda-12.9: Make more declarative

Make use of more vars and declarative environment variables.
This gets rid of some duplication, and will help us break up
the long pipeline step in subsequent commits.

We stop modifying PATH and instead use CUDACXX and
CUDA_TOOLKIT_ROOT for the CUDA compiler detection.
The cmake config now appears to need these settings
to find the compiler (it ftbfs otherwise), and having
nvcc in the PATH no longer appears necessary.

Also gets rid of comments that are (IMO) redundant or otherwise
historical and not necessary for package maintenance.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* tritonserver-cuda-12.9: Use working-directory instead of cd'ing

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* tritonserver-cuda-12.9: Split configure, warning hack and build steps

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* tritonserver-cuda-12.9: pin to cmake-3 to fix ftbfs

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* tritonserver-cuda-12.9: whitespace fix

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* tritonserver-cuda-12.9: drop tensorrt configuration

It doesn't appear to be used, likely because TENSORRT
is configured off.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* tritonserver-cuda-12.9: make comments more concise

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* tritonserver-cuda-12.9: remove hack that drops -Werror

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* py3-torch-tensorrt-cuda-12.9: Adapt to the new llvm packaging

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

* tritonserver-backend-pytorch-cuda-12.9: Drop cuda-architectures-override.patch

No longer applies, but is no longer needed.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>

---------

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>