cloud-init: Replace dep on gptfdisk with sgdisk
A `grep` of cloud-init's source shows that the only command it uses out of
the gptfdisk package is `sgdisk`. The gptfdisk config in enterprise included
`sgdisk` in the `gptfdisk` apk. But that's a duplicate config with one in
wolfi, so we're removing it. The `gptfdisk` config in wolfi has `sfdisk` as
a separate package. While the `gptfdisk` there depends on `sfdisk`, only
`sfdisk` apk seems to be necessary, so let's tighten the dependency.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>