Delete py3.10-safetensors which exists in wolfi (#27494)
py3-safetensors exists in wolfi and is multi-versioned, so it also provides
a py3.10-safetensors. While wolfi is up-to-date at 0.6.0, this version in
enterprise is stuck on 0.5.3 for some reason, so we don't actually have
duplicate apks at the moment.

The enterprise config did have better testing, so I've ported those back
to wolfi:

https://github.com/wolfi-dev/os/pull/59235

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>