actions/vms-test: build compliance benchmarks
The complianceascode-content repository is an internal fork of the upstream
ComplianceAsCode/content repository. It contains everything to build compliance
benchmarks in SCAP format for various Linux distributions. Our internal fork
can be used to develop compliance benchmarks for Chainguard VMs. For now, it
contains a draft of a CIS Benchmark.

Build the compliance benchmarks for Chainguard VMs as part of the vms-test
action. A later commit will add an actual test using the content to scan images
for their grade of compliance.

The compliance scan is opt-in while the test is in development. The plan is to
eventually generate compliance reports for all images on every nightly build.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>