nvidia-libnvjitlink-12.9: Declare a fork of LicenseRef-NVIDIA-CUDA (#3035)
When this package is co-installed with other CUDA packages, apko aborts
because it finds that it has different text in its LICENSE file than other
packages that use "license: LicenseRef-NVIDIA-CUDA", preventing it from
consolidating them. I found that this is because there are 2 additional
sections only in the libnvjitlink LICENSE file:

 23. Portions of support for math operations on 128-bit floating-point data type in compiler
      were derived from SLEEF library which is subject to the following license:

     [... Boost License Text ...]

 24. Portions of support for math operations on 128-bit floating-point data type
     in compiler were derived from SoftFloat library which is subject to the following license:

     [... SoftFloat License Text ...]

I'm not sure if this is a refreshed version of the NVIDIA CUDA License
itself. It still says "Last updated: January 12, 2024", as it did previously.
And this is the only package I've ran into with these new sections. So, to
avoid the conflict, let's just fork a package-specific name for this license.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>