build-publish-*: add workflow input for optional compliance scanning
Introduce a new input "with-compliance-scan" to the build and publish workflows
that gets passed through to the same input of the vms-test action. By using it
for manual workflow dispatch, a compliance scan can be triggered. The input
defaults to false, so all automated workflows on push and on schedule should
remain unaffected.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>