vms-test: add compliance/reports test to run compliance scans
Add a new test "compliance/reports" to run scans using OpenSCAP and generate
compliance reports. The objective of the test is to generate reports only. It
does not fail on non-compliance. We may change that in the future when running
tests on pre-hardened images.

For now, the test runs a compliance scan against a CIS Benchmark Level 1. The
test is generic enough that other compliance scans can be performed, if a
suitable SCAP Security Guide (SSG) and profile are present.

The test is a no-op and skips execution if no SSG is present. An SSG can be
provided by adding a "ssg-chainguard-ds.xml.gz" file to the test source's ssg/
directory. We can think of a better mechanism to provide assets to tests once
we have more uses cases for it. For now, this one-file approach works well with
the testing harness and has shown to be convenient for manual testing as well.

As committed, the test will skip execution in all environments and on all
images. A later commit will modify the GitHub workflows to add the SSG where
desired.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>