actions/vms-test: prepare compliance/reports test for running
Prepare (or arm) the compliance/reports test for running by adding the file
with the compressed SCAP Security Guide to the ssg/ directory that gets
embedded during the build. This does not mean the test has to run--this is
still up to the test config--but if it does run, it will actually perform the
scan instead of skipping.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>