aws-*: run compliance test group for all aws-* variants
Run the compliance test group on all images for Amazon EC2. For now, this runs
the compliance/reports test that generates compliance reports. The test is a
no-op on all runs, unless manual invocation of the build-publish-aws workflow
sets the flag to actually run compliance scanning. Once proven to work, this
flag can be toggled to always run actual compliance scanning.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>