arangodb-3.12: populate log file with package (#28201)
* arangodb: populate log file with package

this is a required file at image level and without this we run into the
following error.

```bash
cat: /var/log/arangodb3/arangod.log: No such file or directory
```

* remove logging to stdout

when we built the image again then we found out that we're not getting
any logs to stdout, fall back to using the file approach that we had
before.

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>

* remove creation of arangod.log file as well

given we're not writing to stdout; this is not required as well

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>

---------

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>