cilium-envoy-fips-1.16: port to FIPS 140-3 boringssl
Port cilium-envoy-fips-1.16 based on envoy-fips-1.33 to FIPS 140-3
boringssl build. Remove old genrule override, and instead use bazel
repository override. Patch envoy version command, such that it can be
used to check the version of the statically linked boringssl. Add
tests that check that envoy's boringssl version & boringssl utility
tooling match the build-time configured version and have a CMVP
certificate.

Changes here, are the ones we already ship in this build of envoy:
- https://github.com/chainguard-dev/enterprise-packages/pull/24748

Note that cilium 1.17, and 1.16 are both using envoy 1.33.4 right now,
hence the patches are the same as in:
- https://github.com/chainguard-dev/enterprise-packages/pull/28368