py3-torchvision-cuda-12.9: enable aarch64 builds (#3204)
This is part of the dependency chain for tritonserver-backend-tensorrtllm:

tritonserver-backend-tensorrtllm-cuda-12.9
 -> py3-nvidia-modelopt-cuda-12.9
   -> py3-torchprofile-cuda-12.9
     -> py3-torchvision-cuda-12.9

We don't currently have plans to build tritonserver for aarch64. However,
torchvision for aarch64 will unblock pytorch images for aarch64 which
I expect we'll want to do soon-ish, and this means we don't have to
restrict py3-nvidia-modelopt and py3-torchprofile to x86-64.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>