py3-tensorrt-cuda-12: add tensorrt-libs package (#3209)
pypi's tensorrt package depends on a tensorrt-libs that provides the
TensorRT .so files. A tensorrt-libs package didn't seem to be necessary
for us because we could just have our py3-tensorrt package depend on the
tensorrt system libraries. But TensorRT apparently has its own library
loader that will fail if the libs aren't found in the python tree or in
`LD_LIBRARY_PATH`.

See https://github.com/chainguard-dev/internal-dev/issues/10372#issuecomment-3128171562 for details.

Setting LD_LIBRARY_PATH everywhere is error-prone, so let's go ahead and
create a -libs package package that depends on and symlinks to the system
libraries.

Because we need to use the build tree that the bindings build leaves behind,
we get rid of the clean step and instead use separate build trees for each
python version.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>