py3-torchvision-cuda-12.9: Use nvidia-cuda-nvcc-gcc-$cudaversion
Switch to the new nvidia-cuda-nvcc-gcc-${{vars.cuda-version}} package to
automatically configure the environment to use the newest version of gcc
supported by the given version of CUDA.

Becuase this config has been copied from older pre-CUDA-12.9 configs,
it has a cargo-culted dep on gcc-12. Using this new nvcc-gcc package
will make it so we just have to store the nvcc version/latest-supported-gcc
mapping in one place, and everyone else will pick it up.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>