configs: update systemd-repart-rootfs (#547)
The systemd-repart-rootfs sub-package of systemd no longer exists under
this name, and hence we are consuming an old build of it. The package
has been split into two filesystem-specific ones when support for ext4
was added (https://github.com/wolfi-dev/os/pull/49716). Consume the new
systemd-repart-rootfs-xfs since for all image builds since all are using
XFS on the root partition.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>