vms-test: do not fail if test does not emit any metrics
Do not fail a test run if a test does not emit any metrics. Instead, check
whether an artifacts file exists before downloading it. Still fail the test if
downloading then actually fails.

The test runners have a heuristic to determine whether a test may be producing
artifacts. The heuristic may produce a false positive, however, if a test
decides at runtime not to produce any artifacts. Keep the heuristic for
efficiency, but do not rely on it exclusively.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>