feat: add RPi images (#529)
This will add the ability to build RPi 5 images.
Added a custom builder as it needs some different steps compared to
regular VM images. Now the makefile accepts the `BUILDER=foo` variable,
so that it will use `foo.yaml` as a builder to create the disk.

Currently two images are generated, one base image and a docker one.

Added to the README steps to build and flash.

Signed-off-by: Luca Di Maio <luca.dimaio1@gmail.com>