vms-test: use the same naming pattern for test artifact files
All test runners (AWS, Azure, GCP, QEMU) used slightly different naming
patterns for the test artifacts files so far. As it seems, by accident. For a
test "reports" in group "compliance" (i.e. "compliance/reports"), the
individual runners would store artifacts with the following names:

    1. AWS: "reports-artifacts.json"
    2. Azure: "reportscomplianceartifacts.json"
    3. GCP: "compliance-reportsartifacts.json"
    4. QEMU: "compliance-reports-artifacts.json"

Unify the naming patterns to the one used in the QEMU runner, for it contains
all identifying information in an easy to read format.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>