pkg: bump test timeouts (#554)
Authenticated go tests have started to timeout and fail. It is
possible it is related to the virtualapk.cgr.dev usage which seems to
be slower than packages.cgr.dev, or it might be due to organic growth
in size of our images.

Do a one-off bump of these timeouts, which should hopefully lead to
more CI check runs passing on the first attempt, without needed a
human to notice this and click retry.

TODO: does this need to go into mono?