build-disks: limit concurrency based on memory needed for VMs
For disk image conversion each VM is allocated up to 4 GiB of RAM (plus
potential in-memory storage of disk images and scratch files). Since an 8 core
runner provides 32 GiB of RAM, the CI job may run out of memory. Limit the
concurrency of the disk image conversion based on the free memory instead of
the number of cores.

Need to see how this works out in practice. Intentionally not using
MemAvailable for the calculation to avoid memory pressure. Eventually, this all
moves to ApkoaaS anyway.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>