build-publish-aws: bump credential lifetime to 2 hours
By default, the ephemeral AWS credentials obtained by the GitHub action are
valid for 1 hour. With the addition of application images, this is not enough
time to import all images before credential expiration and the CI job fails:

    Waiter SnapshotImported failed: An error occurred (RequestExpired): Request has expired.

Increase the credential lifetime to 2 hours, since the number of images roughly
doubled as well and is expected to grow in the near future.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>