cloud-init/25.1.4 package update (#29466)
Bump the cloud-init version to 25.1.4.  This did not happen
automatically because there has not been a release tagged.

At the moment I'm not going to change the updater to `use-tag: true`
as they have done a good job in the past.

For more information, see
See https://discourse.ubuntu.com/t/security-notification-cloud-init-behavior-change-affecting-non-x86-architectures-cve-2024-6174/65268 .