request-4334: add java cacerts and remove bash (#29391)
The dependency on bash is pulling a shell into production images when
the package is added to images via Custom Assembly. This is not
desirable for Danske Bank.

The custom profile script that sets up the the APK repositories is only useful
in dev tags and they always have busybox and bash, so we don't need to add an
explicit dependency for it.