elasticsearch-fips-9: port autoconfiguration to bcfips
This ports autoconfiguration to use bcfks, and automatically sets
default node keystore password. This makes elasticsearch-fips-9 to be
identical drop in replacement of the non-fips versions, with support
support for automatic node configuration, CA generation, Server cert
generation and keystore setup.

This drastically simplified testing, and stock elasticsearch.yml and
for users to deploy it.