amazon-eks-ami, amazon-eks-ami-fips: drop upstream soci-snapshotter units
The upstream Amazon EKS AMI project gained support for AWS' SOCI Snapshotter
containerd plugin [1] that can lazy-load container images. Since the plugin is
not yet packaged for Chainguard OS and the support for it in the upstream
project is still experimental, drop the corresponding systemd units (that would
otherwise cause the system to boot in degraded state).

[1]: https://github.com/awslabs/soci-snapshotter

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>