build-publish-aws: publish if latest build is older than 7 days
Commit 8ca6ba52eec4 ("build-publish-aws: publish to AWS Marketplace once a week only (#501)")
modified the scheduled build-and-publish jobs to only publish images to the AWS
Marketplace once a week, on Mondays. This was done to prevent running into the
low limit of 100 versions per product set by AWS. This decision can quickly
lead to stale images, though, when the next (or even multiple) CI runs fail to
build or qualify new images.

Modify the build-publish-aws workflow once more to publish on Mondays always,
but also any other day when the latest published builds are older than 7 days.
In other words, if a scheduled publishing on Monday fails due to CI failures
that are fixed on Tuesday, the run in the night from Tuesday to Wednesday will
publish images to the Marketplace again. A new set of images will be published
the next Monday.

Since the decision whether to publish is growing more complex, extract it into
a script where it can be tested independently.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>