keycloak: migrate fips, backfill 26.2 (#28922)
Backfill 26.2 for keycloak and 26.2 through 26.3 for keycloak-fips as they would currently be in grace period. 26.1 is technically in grace period for another 7 weeks, but its tests continued to fail after a best effort, so it has been marked as broken in the version stream and should not be backfilled, pending vs creator changes to be deployed to prod-images.