Add TRIAL.md
This is fairly bare bones but gives us something to start with.

So far we have instructions for syncing upstreams and running `stereo
lint` with some caveats around working with git subtrees. As we start
landing automation, this file should become less necessary, as CI should
just guide users to do things correctly.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>